authorization/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
confluent-kafka==2.8.0
cryptography==44.0.0
distlib==0.3.9
dotenv==0.9.9
ecdsa==0.19.0
fastapi==0.115.6
filelock==3.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
passlib==1.7.4
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.29.1
